chkjson: switch to a large inline state machine
This is ugly, but it is also 2x faster.

A very few select functions are not inlined; in my own checks, inlining
them was slower.

There were a few instances while writing this where inlining a function
slowed something down until the next block was inlined; it may be worth
it to un-inline some of this to see if _that_ can speed things up.
Ultimately, every time this happened, the next block that was inlined
more than made up for the difference.

The next commit will address the compacting code, and then a few more
test cases are added.