compact: switch to a large inline state machine
As with the prior commit, this is over 2x faster.

Almost all functions are inlined; this is ugly and duplicative. However,
it is easy to cross verify with chkjson when reading the files side by
side.

The next commit will add a few test cases that get this package to 100%
code coverage.