more minor speed improvements
Surprisingly, moving the isSpace checks to our switch statements more
often speeds things up. This is likely due to the binary search that
switch checks get compiled down to.

This removes isSpace in favor of switches where possible (and inlines it
in the one spot where it is not beneficial to use switch in compact).

This adds back parseStateStack's push; inlining push was not beneficial.