compact: minor optimizations
1.12 apparently loses some compilations niceties (citm got 12% slower!)

This introduces a minor optimization to only increment start where
necessary, as opposed to copying `at` every character loop. This gains
back ~5% of the performance.

There is another oddball performance attempt that isn't universally
applied: rather than checking string comparison equals, we check three
characters individually. The prior strcmp first checks the length of the
string, where we already know they will be equal, so we skip that
comparison check.

This isn't universally applied because I couldn't see a benefit when
benchmarking in the false case.