README: update benchmarks for 1.12
In almost all cases, the stdlib sped up just a (very) litle bit for
itself, and slowed down just a little bit for this library.

It is hard to tell what to blame, but the difference is so minor that it
is not worth it to look into right now.